Reorganising estimators
Reorg, so that more code is shared. Specifically: builds on the
base finite and infinite horizon classes, changing only the update
function as the underlying Q Table stays the same.

Next: Give each Q Estimator an inf and finite implementation.

Tests:
  - test_main.py passes with the new Q Estimator classes. Largely
    copy and pasted.
